This change breakes the binary compatibility of the 4robin.cc files!
The change is that the flat wrappers in the 4robin.cc file would get an array of parameters, take whatever they need from them, cast them to the right parameters and send them to the real function.
This will take off the limit of 12 parametes and/or the ugly huck in call_lowlevel() that allocates space on the stack to make it work.

This change should be widely checked before getting in the trunk.